refactor: 기존의 TokenService를 대체
