+++ src/main/java/com/j9/bestmoments/controller/AuthController.java
@@ -7,7 +7,6 @@ import com.j9.bestmoments.service.OAuthService;
 import com.j9.bestmoments.domain.Member;
 import com.j9.bestmoments.service.MemberService;
 import com.j9.bestmoments.service.TokenService;
-import com.j9.bestmoments.service.Token_Service;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
