+++ src/main/java/com/j9/bestmoments/domain/RefreshToken.java
@@ -4,7 +4,7 @@ import com.j9.bestmoments.constants.TokenExpiration;
 import org.springframework.data.redis.core.RedisHash;
 
 @RedisHash(value = "RefreshToken", timeToLive = TokenExpiration.REFRESH_TOKEN)
-public class RefreshToken extends Token_ {
+public class RefreshToken extends Token {
 
     public RefreshToken(Member member, String token) {
         super(member, token);
