+++ src/main/java/com/j9/bestmoments/repository/Token_Repository.java
@@ -1,11 +1,11 @@
 package com.j9.bestmoments.repository;
 
-import com.j9.bestmoments.domain.Token_;
+import com.j9.bestmoments.domain.Token;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Token_Repository extends CrudRepository<Token_, String> {
+public interface Token_Repository extends CrudRepository<Token, String> {
 
     boolean existsByToken(String token);
